pageから呼び出せるようにプロバイダの読み込み場所を変更
