+++ front/component/modules/auth/SigninContent.tsx
@@ -16,30 +16,13 @@ import usePost from "../../../utility/customhooks/usePost";
 import useIsLogin from "../../../utility/customhooks/useIsLogin";
 
 import { ISignin } from "../../../pages/interface/signin";
-import useToast from "../../../utility/customhooks/useToast";
 
 const SigninContent: React.FC = () => {
   const router = useRouter();
-  const query = router.query;
 
   const apiPost = usePost();
-  const toast = useToast();
   const login = useLogin();
   const [isLogin] = useIsLogin();
-
-  useEffect(() => {
-    console.log(query);
-    if (!query.type) return;
-
-    if (query.type === "register") {
-      toast({
-        text: "greenLifeへようこそ！サインインして早速利用してみましょう！",
-      });
-    }
-    if (query.type === "timeout") {
-      toast({ text: "タイムアウトしました、再度ログインをし直してください。" });
-    }
-  }, [query]);
 
   const [signinUser, setSigninUser] = useState<ISignin>({
     email: "",
