+++ front/pages/_document.js
@@ -19,6 +19,10 @@ export default class MyDocument extends Document {
 
         styles: (
           <>
+            <link
+              href="https://fonts.googleapis.com/css2?family=Bitter:wght@400;600&family=Noto+Sans+JP:wght@400;500&display=swap"
+              rel="stylesheet"
+            />
             {initialProps.styles}
             {sheet.getStyleElement()}
           </>
